+++ doc/CHANGELOG.md
@@ -9,7 +9,8 @@ files use git-show or git-log.
 ## Versions:
 ### vX.Y.Z (WIP):
 - configure.ac added,
-- control keys changed.
+- control keys changed,
+- moveable cursor.
 
 ### v2.0.0 (2018, September 11):
 - vertical && horizontal scroll,